move duplicate tests into evm.rs and keep only the builders per evm chain
